requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
forex-python==1.8
idna==3.4
mysql-connector-python==8.1.0
protobuf==4.21.12
requests==2.31.0
schedule==1.2.1
simplejson==3.19.2
urllib3==2.0.6
